Partial refactoring of the Lea wp rules and FTLR case
- new generic [wp_lea] wp-rule that handles all cases
- simpler proof for the Lea FTLR case, using wp_lea

However, the current statement of wp_lea requires ownership over all
registers, which prevents deriving the existing wp rules as
corollaries (and might be inconvenient for examples).

Hence, for the moment, we keep the proofs for the (many) existing wp
rules as they are.

This is to be addressed later.